requirements.txt
tensorflow
pandas
numpy
matplotlib
flask
jinja2
keras
ipython
scipy
seaborn
scikit-learn
pillow
plotly
mysql.connector
Flask-SQLAlchemy
mysqlclient
pymysql
bcrypt
Werkzeug